% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varFW.R
\name{historicalSim}
\alias{historicalSim}
\title{Historical Simulation}
\usage{
historicalSim(histRet, alpha)
}
\arguments{
\item{histRet}{Vector of historical returns}

\item{alpha}{Alpha value for confidence level}
}
\value{
Max of lowest alpha*100 of histP
}
\description{
Historical Simulation
}
