% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varFW.R
\name{monteCarlo}
\alias{monteCarlo}
\title{Monte Carlo}
\usage{
monteCarlo(mean, sd, n, delta_t)
}
\arguments{
\item{mean}{Sample mean return}

\item{sd}{Sample standard deviation return}

\item{n}{Number of iterations}

\item{delta_t}{Change in time (usually days)}
}
\value{
Vector of simulated returns
}
\description{
Monte Carlo
}
