% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varFW.R
\name{parametric}
\alias{parametric}
\title{Parametric}
\usage{
parametric(mean, sd, alpha, delta_t, ES = FALSE)
}
\arguments{
\item{mean}{Mean return of timeseries data}

\item{sd}{Standard deviation value of timeseries data}

\item{alpha}{Alpha value for confidence level}

\item{delta_t}{Number of days to forecast VaR}
}
\value{
Vector of Adjusted close prices
}
\description{
Parametric
}
