\name{var-package}
\alias{var-package}
\alias{var}
\docType{package}
\title{
\packageTitle{var}
}
\description{
\packageDescription{var}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{var}
\packageIndices{var}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{var}

Maintainer: \packageMaintainer{var}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
