% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varFW.R
\name{VaR}
\alias{VaR}
\title{Value at Risk Wrapper Function}
\usage{
VaR(vLogReturn, alpha = 0.05, n = 1e+06, delta_t = 1,
  method = "parametric", ES = FALSE)
}
\arguments{
\item{vLogReturn}{Vector of Log Returns}

\item{alpha}{Confidence level}

\item{n}{Number of iterations. Default: 1000000}

\item{delta_t}{Time to forecast value at risk. Default: 1}

\item{method}{Method of calculation. Default: parametric. (montecarlo, historical)}

\item{ES}{Expected Shortfall (Conditional VaR). Default: FALSE}
}
\value{
Value at Risk value
}
\description{
Value at Risk Wrapper Function
}
