% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varFW.R
\name{varStats}
\alias{varStats}
\title{Value at Risk Statistics}
\usage{
varStats(calibrationVaR, testMin)
}
\arguments{
\item{calibrationVaR}{Calibrated VaR}

\item{testMin}{Minimum return}
}
\value{
Vector of Calibrated VaR and Minimum Return
}
\description{
Value at Risk Statistics
}
